+++ packages/react-native/types/index.d.ts
@@ -1,4 +1,0 @@
-export { LinkProps, LinkState, ActiveLink } from "./Link";
-export { Active, Block, CuriBase, Curious } from "@curi/react";
-import Link from "./Link";
-export { Link };