+++ packages/react/tests/Active.spec.tsx
@@ -1,21 +1,26 @@
-import 'jest';
-import React from 'react';
-import { shallow } from 'enzyme';
-import InMemory from '@hickory/in-memory';
-import { HickoryLocation } from '@hickory/root';
-import curi, { Response } from '@curi/core';
-import createActiveAddon from '@curi/addon-active';
-import Active from '../src/Active';
+import "jest";
+import React from "react";
+import { mount } from "enzyme";
+import InMemory from "@hickory/in-memory";
+import { HickoryLocation } from "@hickory/root";
+import curi, { Response } from "@curi/core";
+import createActiveAddon from "@curi/addon-active";
+import CuriProvider from "../src/CuriProvider";
+import Active from "../src/Active";
 
-describe('<Active>', () => {
+function render(router, fn) {
+  return mount(<CuriProvider router={router} render={fn} />);
+}
+
+describe("<Active>", () => {
   let history;
   let router;
   const routes = [
-    { name: 'Home', path: '' },
+    { name: "Home", path: "" },
     {
-      name: 'Contact',
-      path: 'contact',
-      children: [{ name: 'Method', path: ':method' }]
+      name: "Contact",
+      path: "contact",
+      children: [{ name: "Method", path: ":method" }]
     }
   ];
 
@@ -26,251 +31,157 @@ describe('<Active>', () => {
     });
   });
 
-  describe('curi and response', () => {
-    function merge(props) {
-      return { ...props, className: 'not-a-test' };
-    }
-    const Test = () => null;
+  describe("no active addon", () => {
+    it('warns if attempting to use in a Curi router without the "active" addon', () => {
+      const router = curi(history, routes);
+      const Test = () => null;
+      function merge(props) {
+        return props;
+      }
 
-    it('can get the router/response values from props', () => {
-      const fakeResponse = {
-        name: 'Home',
-        params: {},
-        partials: []
-      } as Response;
-
-      const wrapper = shallow(
-        <Active name="Home" merge={merge} curi={router} response={fakeResponse}>
-          <Test />
-        </Active>
-      );
-      expect(wrapper.type()).toBe(Test);
-      expect(wrapper.prop('className')).toBe('not-a-test');
-    });
-
-    it('can get the values from context', () => {
-      const fakeResponse = {
-        name: 'Home',
-        params: {},
-        partials: []
-      } as Response;
-
-      const wrapper = shallow(
-        <Active name="Home" merge={merge}>
-          <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
-      expect(wrapper.type()).toBe(Test);
-      expect(wrapper.prop('className')).toBe('not-a-test');
-    });
-
-    it('prefers props over context', () => {
-      const propResponse = {
-        name: 'Home',
-        params: {},
-        partials: []
-      } as Response;
-      const contextResponse = {
-        name: 'House',
-        params: {},
-        partials: []
-      } as Response;
-
-      const wrapper = shallow(
-        <Active name="Home" merge={merge} curi={router} response={propResponse}>
-          <Test />
-        </Active>,
-        { context: { curi: { router, response: contextResponse } } }
-      );
-      expect(wrapper.type()).toBe(Test);
-      expect(wrapper.prop('className')).toBe('not-a-test');
-    });
-
-    it('errors if it cannot access a curi router', () => {
-      const err = console.error;
-      console.error = () => {};
+      const realError = console.error;
+      console.error = jest.fn();
 
       expect(() => {
-        const wrapper = shallow(
+        render(router, () => (
           <Active name="Home" merge={merge}>
             <Test />
           </Active>
-        );
-      }).toThrow();
-
-      console.error = err;
-    });
-  });
-
-  describe('no active addon', () => {
-    it('warns if attempting to use in a Curi router without the "active" addon', () => {
-      const router = curi(history, routes);
-      const fakeResponse = { name: 'Fake', params: {}, partials: [] };
-      const Test = () => null;
-      function merge(props) {
-        return props;
-      }
-
-      expect(() => {
-        shallow(
-          <Active name="Home" merge={merge}>
-            <Test />
-          </Active>,
-          { context: { curi: { router, response: fakeResponse } } }
-        );
+        ));
       }).toThrow(
         'You are attempting to use the "active" prop, but have not included the "active" ' +
-          'addon (curi-addon-active) in your Curi router.'
+          "addon (@curi/addon-active) in your Curi router."
       );
+      console.error = realError;
     });
   });
 
-  describe('children', () => {
-    it('re-renders the children element it is passed', () => {
-      const fakeResponse = { name: 'Fake', params: {}, partials: [] };
+  describe("children", () => {
+    it("re-renders the children element it is passed", () => {
       const Test = () => null;
       function merge(props) {
         return props;
       }
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
-      expect(wrapper.type()).toBe(Test);
+        </Active>
+      ));
+      expect(wrapper.find(Test).exists()).toBe(true);
     });
   });
 
-  describe('merge', () => {
-    it('does not call merge function when not active', () => {
-      const fakeResponse = { name: 'Fake', params: {}, partials: [] };
+  describe("merge", () => {
+    it("does not call merge function when not active", () => {
       const Test = () => null;
       const merge = jest.fn();
 
-      const wrapper = shallow(
-        <Active name="Home" merge={merge}>
+      const wrapper = render(router, () => (
+        <Active name="About" merge={merge}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
+        </Active>
+      ));
       expect(merge.mock.calls.length).toBe(0);
     });
 
-    it('calls merge function when active', () => {
-      const fakeResponse = { name: 'Home', params: {}, partials: [] };
+    it("calls merge function when active", () => {
       const Test = () => null;
       const merge = jest.fn();
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
+        </Active>
+      ));
       expect(merge.mock.calls.length).toBe(1);
     });
 
     it("merges props into children element's props when active", () => {
-      const fakeResponse = { name: 'Home', params: {}, partials: [] };
       const Test = () => null;
       function merge(props) {
-        props.className = 'not-a-test';
+        props.className = "not-a-test";
         return props;
       }
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge}>
           <div className="test" />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
-      expect(wrapper.prop('className')).toBe('not-a-test');
+        </Active>
+      ));
+      const div = wrapper.find("div");
+      expect(div.prop("className")).toBe("not-a-test");
     });
   });
 
-  describe('partial', () => {
-    it('works for partial matches when partial=true', () => {
-      const fakeResponse = {
-        name: 'Method',
-        params: {},
-        partials: ['Contact']
-      };
+  describe("partial", () => {
+    it("works for partial matches when partial=true", done => {
       const Test = () => null;
       function merge(props) {
-        props.className = 'not-a-test';
+        props.className = "not-a-test";
         return props;
       }
-
-      const wrapper = shallow(
-        <Active name="Contact" partial={true} merge={merge}>
-          <div className="test" />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
+      const history = InMemory({ locations: ["/contact/email"] });
+      const router = curi(history, routes, {
+        addons: [createActiveAddon()]
+      });
+      router.respond(
+        () => {
+          const wrapper = render(router, () => (
+            <Active name="Contact" partial={true} merge={merge}>
+              <div className="test" />
+            </Active>
+          ));
+          const div = wrapper.find("div");
+          expect(div.prop("className")).toBe("not-a-test");
+          done();
+        },
+        { once: true }
       );
-      expect(wrapper.prop('className')).toBe('not-a-test');
     });
   });
 
-  describe('extra', () => {
-    it('does nothing when not provided', () => {
-      const fakeResponse = { name: 'Home', params: {}, partials: [] };
+  describe("extra", () => {
+    it("does nothing when not provided", () => {
       const Test = () => null;
       const merge = jest.fn();
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
+        </Active>
+      ));
       expect(merge.mock.calls.length).toBe(1);
     });
 
-    it('passes current location and details to the extra function', () => {
-      const fakeResponse: Response = {
-        name: 'Home',
-        params: {},
-        partials: [],
-        location: { pathname: '/', query: 'test=ing' }
-      } as Response;
+    it("passes current location and details to the extra function", () => {
       const Test = () => null;
       const merge = jest.fn();
-      const details = { query: 'test=ing' };
+      const details = { query: "test=ing" };
       const extra = jest.fn((loc: HickoryLocation, deets: object): boolean => {
-        expect(loc).toBe(fakeResponse.location);
+        //expect(loc).toBe(fakeResponse.location);
         expect(deets).toBe(details);
         return true;
       });
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge} extra={extra} details={details}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
+        </Active>
+      ));
       expect(extra.mock.calls.length).toBe(1);
     });
 
-    it('component is not active if extra returns false', () => {
-      const fakeResponse: Response = {
-        name: 'Home',
-        params: {},
-        partials: [],
-        location: { pathname: '/', query: 'test=ing' }
-      } as Response;
+    it("component is not active if extra returns false", () => {
       const Test = () => null;
       const merge = jest.fn();
       const extra = () => false;
 
-      const wrapper = shallow(
+      const wrapper = render(router, () => (
         <Active name="Home" merge={merge} extra={extra}>
           <Test />
-        </Active>,
-        { context: { curi: { router, response: fakeResponse } } }
-      );
+        </Active>
+      ));
       expect(merge.mock.calls.length).toBe(0);
     });
   });