+++ packages/react/types/Link.d.ts
@@ -1,7 +1,5 @@
 /// <reference types="react" />
 import React from "react";
-import PropTypes from "prop-types";
-import { CuriContext } from "./interface";
 import { CuriRouter, Response } from "@curi/core";
 import { HickoryLocation } from "@hickory/root";
 export interface ActiveLink {
@@ -17,21 +15,13 @@ export interface LinkProps extends React.AnchorHTMLAttributes<HTMLAnchorElement>
     active?: ActiveLink;
     anchor?: React.ReactType;
     target?: string;
-    router?: CuriRouter;
-    response?: Response;
+}
+export interface BaseLinkProps extends LinkProps {
+    router: CuriRouter;
+    response: Response;
 }
 export interface LinkState {
     pathname: string;
 }
-declare class Link extends React.Component<LinkProps, LinkState> {
-    static contextTypes: {
-        curi: PropTypes.Requireable<any>;
-    };
-    clickHandler: (event: React.MouseEvent<HTMLElement>) => void;
-    createPathname(props: LinkProps, context: CuriContext): void;
-    componentWillMount(): void;
-    componentWillReceiveProps(nextProps: LinkProps, nextContext: CuriContext): void;
-    verifyActiveAddon(): void;
-    render(): React.ReactElement<any>;
-}
+declare const Link: (props: LinkProps) => React.ReactElement<any>;
 export default Link;