+++ packages/react/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* User `react-broadcast` to simulate new context API.
 * (Internal) Router's response handler receives `Emitted` object.
 
 ## 1.0.0-beta.19